IT_tBot/requirements.txt
pytelegrambotapi==4.0.0
flask==2.1.0
aiogram==2.24
requests
python-dotenv
